+++ inlining/expected-methods-x86-std-avx2.txt
@@ -4,6 +4,6 @@
 <simdutf8::basic::Utf8Error as core::fmt::Display>::fmt
 <simdutf8::compat::Utf8Error as core::fmt::Debug>::fmt
 <simdutf8::compat::Utf8Error as core::fmt::Display>::fmt
-simdutf8::implementation::algorithm::get_compat_error
+simdutf8::implementation::helpers::get_compat_error
 simdutf8::implementation::x86::validate_utf8_basic
 simdutf8::implementation::x86::validate_utf8_compat
